Changed zoom property for the Coordinate instance from double to int

+++ src/main/java/com/ushahidi/plugins/mapping/maps/providers/yahoo/Tiles.java
@@ -8,7 +8,7 @@ public class Tiles {
 		// Return x, y, z for Yahoo tile column, row, zoom.
 		double x = coordinate.col;
 		double y = Math.pow(2, coordinate.zoom - 1) - coordinate.row - 1;
-		double z = 18 - coordinate.zoom;
+		int z = 18 - coordinate.zoom;
 		return new Coordinate(x, y, z);
 	}
